refactor: 결제 취소 리팩토링
