+++ src/main/java/com/ureca/snac/trade/service/MatchingServiceFacade.java
@@ -36,6 +36,7 @@ public class MatchingServiceFacade {
     private final TradeInitiationService tradeInitiationService;
     private final TradeQueryService tradeQueryService;
     private final TradeProgressService tradeProgressService;
+    private final TradeCancelService tradeCancelService;
     private final BuyFilterService buyFilterService;
     private final AttachmentService attachmentService;
 
@@ -125,7 +126,7 @@ public class MatchingServiceFacade {
         notificationService.notify(tradeDto.getBuyer(), tradeDto);
 
         // 3. 연결되지 않는 다른 거래는 자동 취소
-        List<TradeDto> trades = tradeProgressService.cancelOtherTradesOfCard(tradeDto.getCardId(), tradeId);
+        List<TradeDto> trades = tradeCancelService.cancelOtherTradesOfCard(tradeDto.getCardId(), tradeId);
 
         for (TradeDto dto : trades) {
             CancelTradeDto cancelTradeDto = new CancelTradeDto(dto.getBuyer(), dto);
@@ -172,7 +173,7 @@ public class MatchingServiceFacade {
     // BUY_REQUEST 구매자 기준 취소
     @Transactional
     public void cancelBuyRequestByBuyer(CancelBuyRequest request, String username) {
-        TradeDto tradeDto = tradeProgressService.cancelBuyRequestByBuyerOfCard(request, username);
+        TradeDto tradeDto = tradeCancelService.cancelBuyRequestByBuyerOfCard(request, username);
 
         notificationService.sendCancelNotification(new CancelTradeDto(tradeDto.getSeller(), tradeDto));
     }
@@ -180,7 +181,7 @@ public class MatchingServiceFacade {
     // BUY_REQUEST 판매자 기준 취소
     @Transactional
     public void cancelBuyRequestBySeller(CancelBuyRequest request, String username) {
-        List<TradeDto> trades = tradeProgressService.cancelBuyRequestBySellerOfCard(request, username);
+        List<TradeDto> trades = tradeCancelService.cancelBuyRequestBySellerOfCard(request, username);
 
         for (TradeDto tradeDto : trades) {
             notificationService.sendCancelNotification(new CancelTradeDto(tradeDto.getBuyer(), tradeDto));
@@ -190,29 +191,29 @@ public class MatchingServiceFacade {
     // ACCEPTED 거래 취소 - 구매자
     @Transactional
     public void cancelAcceptedTradeByBuyer(CancelRealTimeTradeRequest request, String username) {
-        TradeDto tradeDto = tradeProgressService.cancelAcceptedTradeByBuyer(request, username);
+        TradeDto tradeDto = tradeCancelService.cancelRealTimeTrade(request.getTradeId(), username, request.getReason());
         notificationService.sendCancelNotification(new CancelTradeDto(tradeDto.getSeller(), tradeDto));
     }
 
     // ACCEPTED 거래 취소 - 판매자
     @Transactional
     public void cancelAcceptedTradeBySeller(CancelRealTimeTradeRequest request, String username) {
-        TradeDto tradeDto = tradeProgressService.cancelAcceptedTradeBySeller(request, username);
+        TradeDto tradeDto = tradeCancelService.cancelRealTimeTrade(request.getTradeId(), username, request.getReason());
         notificationService.sendCancelNotification(new CancelTradeDto(tradeDto.getBuyer(), tradeDto));
     }
 
     @Transactional
     public void cancelPaymentTradeByBuyer(CancelRealTimeTradeRequest request, String username) {
-        tradeProgressService.cancelRealTimeTradeWithRefund(request.getTradeId(), username);
+//        tradeCancelService.cancelRealTimeTradeWithRefund(request.getTradeId(), username);
 
-        TradeDto tradeDto = tradeProgressService.cancelRealTimeTrade(request.getTradeId(), username, request.getReason());
+        TradeDto tradeDto = tradeCancelService.cancelRealTimeTrade(request.getTradeId(), username, request.getReason());
         notificationService.sendCancelNotification(new CancelTradeDto(tradeDto.getSeller(), tradeDto));
     }
 
     @Transactional
     public void cancelPaymentTradeBySeller(CancelRealTimeTradeRequest request, String username) {
-        tradeProgressService.cancelRealTimeTradeWithRefund(request.getTradeId(), username);
-        TradeDto tradeDto = tradeProgressService.cancelRealTimeTrade(request.getTradeId(), username, request.getReason());
+//        tradeCancelService.cancelRealTimeTradeWithRefund(request.getTradeId(), username);
+        TradeDto tradeDto = tradeCancelService.cancelRealTimeTrade(request.getTradeId(), username, request.getReason());
 
         notificationService.sendCancelNotification(new CancelTradeDto(tradeDto.getBuyer(), tradeDto));
     }
