+++ src/main/java/com/ureca/snac/trade/service/TradeCancelServiceImpl.java
@@ -5,11 +5,10 @@ import com.ureca.snac.asset.service.AssetChangedEventFactory;
 import com.ureca.snac.asset.service.AssetHistoryEventPublisher;
 import com.ureca.snac.board.entity.Card;
 import com.ureca.snac.member.Member;
+import com.ureca.snac.trade.controller.request.CancelBuyRequest;
+import com.ureca.snac.trade.dto.TradeDto;
 import com.ureca.snac.trade.entity.*;
-import com.ureca.snac.trade.exception.TradeAlreadyCancelRequestedException;
-import com.ureca.snac.trade.exception.TradeCancelNotFoundException;
-import com.ureca.snac.trade.exception.TradeCancelPermissionDeniedException;
-import com.ureca.snac.trade.exception.TradeInvalidStatusException;
+import com.ureca.snac.trade.exception.*;
 import com.ureca.snac.trade.repository.TradeCancelRepository;
 import com.ureca.snac.trade.repository.TradeRepository;
 import com.ureca.snac.trade.service.interfaces.PenaltyService;
@@ -21,6 +20,11 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
+import java.util.List;
+
+import static com.ureca.snac.trade.entity.CancelReason.NOT_SELECTED;
+import static com.ureca.snac.trade.entity.TradeStatus.BUY_REQUESTED;
+import static com.ureca.snac.trade.entity.TradeStatus.CANCELED;
 
 @Service
 @RequiredArgsConstructor
@@ -194,4 +198,127 @@ public class TradeCancelServiceImpl implements TradeCancelService {
 
         cancel.reject();  // 거래 계속 진행
     }
+
+
+    // 거래는 1대1 매칭이기 때문에 선택되지 않는 다른 거래는 모두 자동 취소 처리 됩니다.
+    @Override
+    @Transactional
+    public List<TradeDto> cancelOtherTradesOfCard(Long cardId, Long acceptedTradeId) {
+        List<Trade> waitingTrades = tradeRepo.findLockedByCardIdAndStatus(cardId, BUY_REQUESTED)
+                .stream()
+                .filter(t -> !t.getId().equals(acceptedTradeId))
+                .toList();
+
+        waitingTrades.forEach(t -> {
+            t.changeStatus(CANCELED);
+            t.changeCancelReason(NOT_SELECTED);
+        });
+
+        return waitingTrades.stream()
+                .map(TradeDto::from)
+                .toList();
+    }
+
+    // 판매자가 수락 상태 전에 판매글을 내리는 경우입니다.
+    // 해당 판매자에게 거래 요청을 보낸 모든 구매자의 거래를 취소 처리합니다.
+    // 취소는 구매자가 작성한 값으로 처리합니다. (CancelReason)
+    @Override
+    @Transactional
+    public List<TradeDto> cancelBuyRequestBySellerOfCard(CancelBuyRequest request, String username) {
+        List<Trade> waitingTrades = tradeRepo.findLockedByCardIdAndStatus(request.getCardId(), BUY_REQUESTED);
+
+        cancelRepo.deleteById(request.getCardId());
+
+        waitingTrades.forEach(t -> {
+            t.changeStatus(CANCELED);
+            t.changeCancelReason(request.getReason());
+        });
+
+        return waitingTrades.stream()
+                .map(TradeDto::from)
+                .toList();
+    }
+
+
+    // 구매자가 구매 요청을 보낸 후 수락 단계를 넘어가기 전에 스스로 거래를 취소한 경우입니다.
+    // 판매자에게 해당 요청은 취소되었다는 알림이 전송됩니다.
+    @Override
+    @Transactional
+    public TradeDto cancelBuyRequestByBuyerOfCard(CancelBuyRequest request, String username) {
+        Member member = tradeSupport.findMember(username);
+
+        Trade trade = tradeRepo.findLockedByCardIdAndBuyer(request.getCardId(), member)
+                .orElseThrow(TradeNotFoundException::new);
+
+        trade.cancel(member);
+        trade.changeCancelReason(request.getReason());
+
+        return TradeDto.from(trade);
+    }
+
+    // ACCEPT, PAYMENT 단계의 취소를 처리하는 메서드입니다.
+
+    @Override
+    @Transactional
+    public TradeDto cancelRealTimeTrade(Long tradeId, String username, CancelReason reason) {
+        Member member = tradeSupport.findMember(username);
+        Trade trade = tradeSupport.findLockedTrade(tradeId);
+        Card card = tradeSupport.findLockedCard(trade.getCardId());
+
+        if (trade.getStatus() == TradeStatus.PAYMENT_CONFIRMED) {
+            refundToBuyerAndPublishEvent(trade, card, trade.getBuyer());
+        }
+
+        cancelRepo.deleteById(trade.getCardId());
+
+        trade.cancel(member);
+        trade.changeCancelReason(reason);
+
+        return TradeDto.from(trade);
+    }
+
+    @Override
+    @Transactional
+    public TradeDto cancelRealTimeTradeWithRefund(Long tradeId, String username) {
+        Trade trade = tradeSupport.findLockedTrade(tradeId);
+        Member member = tradeSupport.findMember(username);
+
+        if (trade.getPriceGb() - trade.getPoint() > 0) {
+            walletService.depositMoney(trade.getBuyer().getId(), trade.getPriceGb() - trade.getPoint());
+        }
+
+        if (trade.getPoint() > 0) {
+            walletService.depositPoint(trade.getBuyer().getId(), trade.getPoint());
+        }
+
+        return TradeDto.from(trade);
+    }
+
+
+//    @Override
+//    @Transactional
+//    public TradeDto cancelAcceptedTradeByBuyer(CancelRealTimeTradeRequest cancelRealTimeTradeRequest, String username) {
+//        Member member = tradeSupport.findMember(username);
+//        Trade trade = tradeSupport.findLockedTrade(cancelRealTimeTradeRequest.getTradeId());
+//
+//        cancelRepo.deleteById(trade.getCardId());
+//        trade.cancel(member);
+//        trade.changeCancelReason(BUYER_CHANGE_MIND);
+//
+//        return TradeDto.from(trade);
+//    }
+//
+//    @Override
+//    @Transactional
+//    public TradeDto cancelAcceptedTradeBySeller(CancelRealTimeTradeRequest cancelRealTimeTradeRequest, String username) {
+//        Member member = tradeSupport.findMember(username);
+//        Trade trade = tradeSupport.findLockedTrade(cancelRealTimeTradeRequest.getTradeId());
+//
+//        cancelRepo.deleteById(trade.getCardId());
+//
+//        trade.cancel(member);
+//        trade.changeCancelReason(SELLER_CHANGE_MIND);
+//
+//        return TradeDto.from(trade);
+//    }
 }
