+++ src/main/java/com/ureca/snac/trade/service/interfaces/TradeProgressService.java
@@ -26,17 +26,17 @@ public interface TradeProgressService {
      */
     Long confirmTrade(Long tradeId, String username);
 
-    List<TradeDto> cancelOtherTradesOfCard(Long cardId, Long acceptedTradeId);
-
-    TradeDto cancelBuyRequestByBuyerOfCard(CancelBuyRequest request, String username);
-
-    List<TradeDto> cancelBuyRequestBySellerOfCard(CancelBuyRequest request, String username);
-
-    TradeDto cancelAcceptedTradeByBuyer(CancelRealTimeTradeRequest cancelRealTimeTradeRequest, String username);
-
-    TradeDto cancelAcceptedTradeBySeller(CancelRealTimeTradeRequest cancelRealTimeTradeRequest, String username);
-
-    TradeDto cancelRealTimeTrade(Long tradeId, String username, CancelReason reason);
-
-    TradeDto cancelRealTimeTradeWithRefund(Long tradeId, String username);
+//    List<TradeDto> cancelOtherTradesOfCard(Long cardId, Long acceptedTradeId);
+//
+//    TradeDto cancelBuyRequestByBuyerOfCard(CancelBuyRequest request, String username);
+//
+//    List<TradeDto> cancelBuyRequestBySellerOfCard(CancelBuyRequest request, String username);
+//
+//    TradeDto cancelAcceptedTradeByBuyer(CancelRealTimeTradeRequest cancelRealTimeTradeRequest, String username);
+//
+//    TradeDto cancelAcceptedTradeBySeller(CancelRealTimeTradeRequest cancelRealTimeTradeRequest, String username);
+//
+//    TradeDto cancelRealTimeTrade(Long tradeId, String username, CancelReason reason);
+//
+//    TradeDto cancelRealTimeTradeWithRefund(Long tradeId, String username);
 }
